refactor: move factorization to lyco_helpers, separate calc_updown for kohya and kb
